// apps/api/src/regulations/enums/deduction-type.enum.ts
export enum DeductionType {
  ATHLETE_FALL = 'athlete_fall',
  MAJOR_ATHLETE_FALL = 'major_athlete_fall',
  BUILDING_BOBBLE = 'building_bobble',
  BUILDING_FALL = 'building_fall',
  MAJOR_BUILDING_FALL = 'major_building_fall',
  LEGALITY_INFRACTIONS = 'legality_infractions',
  SKILL_OUT_OF_LEVEL = 'skill_out_of_level',
  TIME_LIMIT_VIOLATIONS = 'time_limit_violations',
  BOUNDARY_VIOLATIONS = 'boundary_violations',
}